Map stuff and more hunt setup
